requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic>=2.8.0
google-generativeai==0.3.1
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
google-cloud-storage>=2.10.0
anthropic>=0.7.0
cryptography>=41.0.0
